functions/requirements.txt
# Function dependencies, for example:
# package>=version

tensorflow
google-cloud-storage